requirements.txt
diffusers>=0.26.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
tqdm>=4.66.0
accelerate>=0.24.0
huggingface-hub>=0.19.0 